Create separate expression selector sheet widget
